+++ packages/admin/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @byline/admin
 
+## 1.6.2
+
+### Patch Changes
+
+- Bundle analyzer and updated cli/manifest deps to include sharp.
+- Updated dependencies
+  - @byline/auth@1.6.2
+  - @byline/core@1.6.2
+
 ## 1.6.1
 
 ### Patch Changes