+++ packages/host-tanstack-start/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @byline/host-tanstack-start
 
+## 1.6.2
+
+### Patch Changes
+
+- Bundle analyzer and updated cli/manifest deps to include sharp.
+- Updated dependencies
+  - @byline/admin@1.6.2
+  - @byline/auth@1.6.2
+  - @byline/client@1.6.2
+  - @byline/core@1.6.2
+  - @byline/ui@1.6.2
+
 ## 1.6.1
 
 ### Patch Changes