+++ packages/storage-s3/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @byline/storage-s3
 
+## 1.6.2
+
+### Patch Changes
+
+- Bundle analyzer and updated cli/manifest deps to include sharp.
+- Updated dependencies
+  - @byline/core@1.6.2
+
 ## 1.6.1
 
 ### Patch Changes